requirements.txt
PySide6>=6.7,<7
loguru>=0.7.2,<1
requests>=2.32.3,<3
Pillow>=11.0.0,<12
modal>=0.64.232,<1
pyyaml>=6.0.2,<7